requirements.txt
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.3
easyocr==1.7.2
et-xmlfile==2.0.0
exceptiongroup==1.3.0
fastapi==0.115.2
filelock==3.20.0
fsspec==2025.10.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.27.2
huggingface-hub==0.26.2
idna==3.11
ImageIO==2.37.2
Jinja2==3.1.6
langdetect==1.0.9
lazy_loader==0.4
lxml==6.0.2
MarkupSafe==3.0.3
mpmath==1.3.0
narwhals==2.12.0
networkx==3.4.2
ninja==1.13.0
numpy==2.2.6
opencv-python-headless==4.12.0.88
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
pdfminer.six==20251107
pdfplumber==0.11.8
pillow==12.0.0
plotly==6.5.0
psutil==6.1.0
pyclipper==1.3.0.post6
pycparser==2.23
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
pyodbc==5.2.0
pypdfium2==5.0.0
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.0.1
python-multipart==0.0.10
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
scikit-image==0.25.2
scipy==1.15.3
shapely==2.1.2
six==1.17.0
sniffio==1.3.1
starlette==0.40.0
sympy==1.14.0
tifffile==2025.5.10
torch==2.9.1
torchvision==0.24.1
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.32.0
waitress==3.0.2
watchfiles==1.1.1
websockets==15.0.1
xlrd==2.0.2
